extract SafeDisposableFrame to its own file
