+++ src/gui/frames/__init__.py
@@ -3,3 +3,4 @@ from .frame_menu import *
 from .frame_profile_editor import *
 from .frame_profile_switcher import *
 from .safe_disposable_frame import *
+from .safe_disposable_scrollable_frame import *
